+++ app/Http/Controllers/CompanyController.php
@@ -107,7 +107,7 @@ class CompanyController extends Controller
 
         $company = Auth::user();
         if($req->logo){
-            if ($company->profile_picture && Storage::disk('public')->exists($company->profile_picture)) {
+            if ($company->profile_picture && $company->profile_picture != 'default/profile_picture.jpg' && Storage::disk('public')->exists($company->profile_picture)) {
                 $test = Storage::disk('public')->delete($company->profile_picture);
             }
